chore: Moves InfoTooltip to the components folder
