+++ superset-frontend/src/common/components/common.stories.tsx
@@ -22,11 +22,11 @@ import { withKnobs, boolean, select } from '@storybook/addon-knobs';
 import Button from 'src/components/Button';
 import { CronPicker, CronError } from 'src/components/CronPicker';
 import Modal from 'src/components/Modal';
+import InfoTooltip from 'src/components/InfoTooltip';
 import Tabs, { EditableTabs } from './Tabs';
 import { Tooltip as AntdTooltip } from './Tooltip';
 import { Menu, Input, Divider } from '.';
 import { Dropdown } from './Dropdown';
-import InfoTooltip from './InfoTooltip';
 
 export default {
   title: 'Common components',
